+++ test/telegraf.js
@@ -1,5 +1,5 @@
 const test = require('ava')
-const { Telegraf, session, Scenes } = require('../')
+const { Telegraf, session } = require('../')
 
 function createBot (...args) {
   const bot = new Telegraf(...args)
@@ -155,61 +155,4 @@ test('should respect webhookReply runtime change (per request)', async (t) => {
   await t.throwsAsync(bot.handleUpdate({ message: BaseTextMessage }, res))
   t.true(res.writableEnded)
   t.is(res.body, undefined)
-})
-
-test('should execute enter middleware in scene', (t) => {
-  const bot = createBot()
-  const scene = new Scenes.BaseScene('hello')
-  scene.enter((ctx) => t.pass())
-  const stage = new Scenes.Stage([scene])
-  stage.use((ctx) => ctx.scene.enter('hello'))
-  bot.use(session({ getSessionKey: (ctx) => '1' }))
-  bot.use(stage)
-  return bot.handleUpdate({ message: BaseTextMessage })
-})
-
-test('should execute enter middleware in wizard scene', (t) => {
-  const bot = createBot()
-  const scene = new Scenes.WizardScene('hello', [])
-  scene.enter((ctx) => t.pass())
-  const stage = new Scenes.Stage([scene])
-  stage.use((ctx) => ctx.scene.enter('hello'))
-  bot.use(session({ getSessionKey: (ctx) => '1' }))
-  bot.use(stage)
-  return bot.handleUpdate({ message: BaseTextMessage })
-})
-
-test('should execute first step in wizard scene on enter', (t) => {
-  const bot = createBot()
-  const scene = new Scenes.WizardScene(
-    'hello',
-    (ctx) => {
-      t.pass()
-    }
-  )
-  const stage = new Scenes.Stage([scene])
-  stage.use((ctx) => ctx.scene.enter('hello'))
-  bot.use(session({ getSessionKey: (ctx) => '1' }))
-  bot.use(stage)
-  return bot.handleUpdate({ message: BaseTextMessage })
-})
-
-test('should execute both enter middleware and first step in wizard scene on enter', (t) => {
-  t.plan(2)
-  const bot = createBot()
-  const scene = new Scenes.WizardScene(
-    'hello',
-    (ctx) => {
-      t.pass()
-    }
-  )
-  scene.enter((ctx, next) => {
-    t.pass()
-    return next()
-  })
-  const stage = new Scenes.Stage([scene])
-  stage.use((ctx) => ctx.scene.enter('hello'))
-  bot.use(session({ getSessionKey: (ctx) => '1' }))
-  bot.use(stage)
-  return bot.handleUpdate({ message: BaseTextMessage })
 })
